Improve the cleanup for the OSS tests

+++ src/acceptance/app/custom_metric_test.go
@@ -3,6 +3,8 @@ package app_test
 import (
 	"acceptance/config"
 	. "acceptance/helpers"
+	"fmt"
+	"os"
 
 	"github.com/cloudfoundry-incubator/cf-test-helpers/cf"
 	"github.com/cloudfoundry-incubator/cf-test-helpers/generator"
@@ -56,13 +58,16 @@ var _ = Describe("AutoScaler custom metrics policy", func() {
 			close(doneChan)
 			Eventually(doneAcceptChan, 30*time.Second).Should(Receive())
 
-			DeletePolicy(appName, appGUID)
-			if !cfg.IsServiceOfferingEnabled() {
-				DeleteCustomMetricCred(appGUID)
+			if os.Getenv("SKIP_TEARDOWN") == "true" {
+				fmt.Println("Skipping Teardown...")
+			} else {
+				DeletePolicy(appName, appGUID)
+				if !cfg.IsServiceOfferingEnabled() {
+					DeleteCustomMetricCred(appGUID)
+				}
+
+				Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.DefaultTimeoutDuration())).To(Exit(0))
 			}
-
-			Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.DefaultTimeoutDuration())).To(Exit(0))
-
 		})
 
 		Context("when test_metric is less than scaling in threshold", func() {
